pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-policy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gym-pusht==0.1.6",
    "zarr==2.18.3",
    "wandb==0.24.1",
    "diffusers>=0.36.0",
    "tqdm>=4.67.2",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "pymunk==6.6.0",
    "sk-video>=1.1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["diffusion_policy*"]
